Fix g++ warning msg
git-svn-id: svn://irodssvn.ucsd.edu/trunk@3937 ea137bd6-aa86-47b6-b90d-b101d9f19a90

+++ iRODS/server/api/src/rsChkNVPathPerm.c
@@ -157,7 +157,8 @@ _rsChkNVPathPerm (rsComm_t *rsComm, fileOpenInp_t *chkNVPathPermInp)
 	return (SYS_NO_PATH_PERMISSION);
     }
 	    
-    if (sysUid != myFileStat.st_uid && (myFileStat.st_mode & S_IWOTH) == 0) {
+    if (sysUid != (int) myFileStat.st_uid && 
+      (myFileStat.st_mode & S_IWOTH) == 0) {
 	return (SYS_NO_PATH_PERMISSION);
     } else {
         return (0);
